Refactor parameter handling and consolidate test cases across modules
